feat: refactor at element (serial and parallel blocks placed correctly)
